Use comment colors for NonText group
The 'list' option uses the NonText group for highlighting things like
tabs, spaces at the end of a line, etc. When the 'list' option is set,
it's nice to actually be able to see the characters that are
highlighted, so change the NonText group from being invisible to
appearing as a comment.